Add Webpack
Add Webpack to bundle all the javascript files into one bundle.js, which
can be included in the html page.

Remove dist from the .gitignore, so that the index.html can be checked
in. The compiled main.js will also be committed, so the project can be
run out of the box (in case reviewers have issues building the code from
source).