Add a dampening factor on collision
Colliding with a wall will absorb some of the velocity from the ball.
Apply a 'dampening factor' on collision to slow the ball down, so it
doesn't bounce forever.

As part of this change, fix a bug where the ball would end up bouncing
infinitely towards the end, because collisions were detected from the
center of the ball (rather than the edges).